requirements.txt
openai~=0.27.2
numpy~=1.24.2
sacrebleu~=2.3.1
backoff~=2.2.1
blobfile~=2.0.1
lz4~=4.3.2
pydantic~=1.10.6
pyzstd~=0.15.4
tqdm~=4.65.0
PyYAML~=6.0
jsonlines~=3.1.0
pandas~=1.5.3
datasets~=2.10.1
jsonlines~=3.1.0
pandas~=1.5.3
datasets~=2.10.1
requests~=2.28.1
boto3~=1.26.104
urllib3~=1.26.15
loguru~=0.6.0
openpyxl~=3.1.2
PyYAML~=6.0
protobuf==3.19.4
grpcio==1.47.0
grpcio-tools==1.44.0
sentence-transformers~=2.2.2
zhon~=1.1.5
Flask~=2.2.3
gevent~=22.10.2
wudao~=1.3.0